refactor: controller service pattern for logout api
