--- src/server/api/logout/index.ts
@@ -1,8 +1,0 @@
-import Express from 'express'
-import { logOut } from './handlers'
-
-const router = Express.Router()
-
-router.get('/', logOut)
-
-export = router
