+++ src/server/controllers/interfaces/LogoutControllerInterface.ts
@@ -1,0 +1,5 @@
+import Express from 'express'
+
+export interface LogoutControllerInterface {
+  logOut(req: Express.Request, res: Express.Response): void
+}
