+++ src/server/inversify.config.ts
@@ -25,6 +25,7 @@ import { RotatingLinksController } from './controllers/RotatingLinksController'
 import { SentryController } from './controllers/SentryController'
 import { LoginController } from './controllers/LoginController'
 import { AuthService } from './services/AuthService'
+import { LogoutController } from './controllers/LogoutController'
 
 function bindIfUnbound<T>(
   dependencyId: symbol,
@@ -55,6 +56,7 @@ export default () => {
   bindIfUnbound(DependencyIds.sentryController, SentryController)
   bindIfUnbound(DependencyIds.loginController, LoginController)
   bindIfUnbound(DependencyIds.authService, AuthService)
+  bindIfUnbound(DependencyIds.logoutController, LogoutController)
 
   container.bind(DependencyIds.s3Bucket).toConstantValue(s3Bucket)
